+++ src/test/scala/edu/ie3/simona/test/common/ConfigTestData.scala
@@ -9,6 +9,8 @@ package edu.ie3.simona.test.common
 import com.typesafe.config.{Config, ConfigFactory}
 import edu.ie3.simona.config.SimonaConfig
 
+import java.time.ZonedDateTime
+
 /** Simple (empty) configuration data. Furthermore, it would make sense to
   * implement another class which reads a config and provides config based
   * values in the future.
@@ -151,4 +153,5 @@ trait ConfigTestData {
   )
   protected val simonaConfig: SimonaConfig = SimonaConfig(typesafeConfig)
 
+  protected val startTime: ZonedDateTime = simonaConfig.simona.time.simStartTime
 }